fix: Use Double instead of Float and rename package

+++ struckoutCameraApp/app/src/main/java/com/taichi765/struckoutCameraApp/camera/types/WorldDirection.kt
@@ -1,3 +1,3 @@
 package com.taichi765.struckoutCameraApp.camera.types
 
-data class WorldDirection(val x: Float, val y: Float, val z: Float)
+data class WorldDirection(val x: Double, val y: Double, val z: Double)
